Added fixes for the tests  - Fix tests to use the create space path instead of static path  - Move Decompressed size header test to its own test, out of RestTests category
